update redis pool usage and cleanup
