+++ explorer/redis.go
@@ -8,25 +8,9 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-var pool *redis.Pool
-
-func InitRedisPool() {
-	pool = &redis.Pool{
-		MaxIdle:   10,
-		MaxActive: 10,
-		Dial: func() (redis.Conn, error) {
-			conn, err := redis.Dial("tcp", "localhost:6379")
-			if err != nil {
-				log.Error().Err(errors.Wrap(err, "ERROR: fail init redis")).Msg("")
-			}
-			return conn, err
-		},
-	}
-}
-
-func SetRedisKey(key string, val []byte, expiration uint64) error {
+func (a *App) SetRedisKey(key string, val []byte, expiration uint64) error {
 	// get conn and put back when exit from method
-	conn := pool.Get()
+	conn := a.redis.Get()
 	defer conn.Close()
 
 	_, err := conn.Do("SET", key, val, "EX", expiration)
@@ -38,9 +22,9 @@ func SetRedisKey(key string, val []byte, expiration uint64) error {
 	return nil
 }
 
-func GetRedisKey(key string) (string, error) {
+func (a *App) GetRedisKey(key string) (string, error) {
 	// get conn and put back when exit from method
-	conn := pool.Get()
+	conn := a.redis.Get()
 	defer conn.Close()
 
 	s, err := redis.String(conn.Do("GET", key))
